moving api -> apis and setting multigroup: true
