+++ PROJECT
@@ -1,6 +1,7 @@
 domain: cluster.x-k8s.io
 layout: go.kubebuilder.io/v3
 projectName: cluster-api-provider-nested
+multigroup: true
 repo: sigs.k8s.io/cluster-api-provider-nested
 resources:
 - group: controlplane
